Update methods to support pointer receiver
`(*metainfo.Hash).HexString()` currently uses a value receiver, which allows for chaining method calls (e.g. `.Hash().HexString()`). This is nice, but I've been doing some testing with pointer receiver methods -- such methods disallow can't be called on immediately returned values.

To maintain compatibility with both pointer and value receiver methods, a local variable needs to be created. This shouldn't be a big deal, but it helps a lot with code refactoring.