Pool readers and create default constructor
I was hesitant to make this request before, but it now seems like a pretty obvious choice as the `reader` struct begins to get more fields. Torrents with thousands of files would easily benefit from this. On the hand, if a user only create a single reader for a torrent -- there would likely be no benefit.

But the added bonus is that it goes along nicely with constructors! The constructor here is obviously `newReader`